School routes organized into separate controller & service
