--- src/api-v2/v2.controller.ts
@@ -1,9 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { V2Service } from './v2.service';
-
-@Controller('api/v2')
-export class V2Controller {
-  constructor(
-    private service: V2Service
-  ) {}
-}
